requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
Appium-Python-Client==2.2.0
argparse3==0.0.0
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
azure-core==1.26.1
azure-cosmos==4.3.0
Babel==2.9.1
beautifulsoup4==4.11.1
blis==0.7.9
build==0.10.0
CacheControl==0.12.11
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
cleo==2.0.1
click==8.1.3
colorama==0.4.3
confection==0.0.4
crashtest==0.4.1
cryptography==36.0.2
cssselect==1.2.0
cuttpy==1.1.1
cymem==2.0.7
decorator==5.1.1
distlib==0.3.6
Django==4.1.7
dnspython==2.3.0
dulwich==0.21.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
es-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.5.0/es_core_news_lg-3.5.0-py3-none-any.whl
es-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/es_core_news_sm-3.4.0/es_core_news_sm-3.4.0-py3-none-any.whl
et-xmlfile==1.0.1
filelock==3.9.0
flatten-json==0.1.13
frozenlist==1.3.3
google==3.0.0
h11==0.13.0
html5lib==1.1
idna==3.4
importlib-metadata==4.13.0
installer==0.7.0
isodate==0.6.1
jaraco.classes==3.2.3
jdcal==1.4.1
Jinja2==3.1.2
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.17.3
keyring==23.13.1
langcodes==3.3.0
lockfile==0.12.2
logging42==0.0.8
loguru==0.6.0
lxml==4.9.1
magic-filter==1.0.9
MarkupSafe==2.1.2
mongo==0.2.0
more-itertools==9.0.0
msgpack==1.0.4
msrest==0.7.1
multidict==6.0.4
murmurhash==1.0.9
newsapi==0.1.1
newsapi-python==0.2.7
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.2
openpyxl==3.0.3
outcome==1.1.0
packaging==22.0
pandas==1.5.3
pathy==0.10.1
pdfkit==1.0.0
pexpect==4.8.0
pi==0.1.2
Pillow==9.4.0
pkginfo==1.9.4
platformdirs==2.6.2
poetry==1.4.1
poetry-core==1.5.2
poetry-dotenv==0.3.0
poetry-plugin-export==1.3.0
preshed==3.0.8
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
pymongo==4.3.3
pynput==1.6.8
pyobjc-core==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-Quartz==9.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gitlab==2.2.0
pytz==2022.7.1
random2==1.0.1
rapidfuzz==2.13.7
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
scikit-learn==1.2.1
scipy==1.10.0
selenium==4.4.3
shellingham==1.5.0
six==1.16.0
smart-open==6.3.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
textblob==0.17.1
thinc==8.1.9
threadpoolctl==3.1.0
tk==0.1.0
tkhtmlview==0.1.3
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.65.0
trio==0.20.0
trio-websocket==0.9.2
trove-classifiers==2022.12.22
tweepy==4.13.0
twilio==7.17.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.13
validators==0.20.0
virtualenv==20.17.1
wasabi==1.1.1
webdriver-manager==3.8.5
webencodings==0.5.1
wget==3.2
wkhtmltopdf==0.2
wsproto==1.1.0
xattr==0.10.1
yarl==1.8.2
zipp==3.11.0
